+++ src/main/java/com/workat/common/exception/ForbiddenException.java
@@ -1,0 +1,16 @@
+package com.workat.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+import com.workat.common.exception.base.BusinessException;
+
+public class ForbiddenException extends BusinessException {
+
+	public ForbiddenException(String msg) {
+		super(HttpStatus.FORBIDDEN, msg);
+	}
+
+	public ForbiddenException(Throwable t) {
+		super(HttpStatus.FORBIDDEN, t);
+	}
+}
