--- src/main/java/com/workat/exception/ConflictException.java
@@ -1,16 +1,0 @@
-package com.workat.exception;
-
-import org.springframework.http.HttpStatus;
-
-import com.workat.exception.base.BusinessException;
-
-public class ConflictException extends BusinessException {
-
-	public ConflictException(String msg) {
-		super(HttpStatus.CONFLICT, msg);
-	}
-
-	public ConflictException(Throwable t) {
-		super(HttpStatus.CONFLICT, t);
-	}
-}
