--- src/main/java/com/workat/exception/ForbiddenException.java
@@ -1,16 +1,0 @@
-package com.workat.exception;
-
-import org.springframework.http.HttpStatus;
-
-import com.workat.exception.base.BusinessException;
-
-public class ForbiddenException extends BusinessException {
-
-	public ForbiddenException(String msg) {
-		super(HttpStatus.FORBIDDEN, msg);
-	}
-
-	public ForbiddenException(Throwable t) {
-		super(HttpStatus.FORBIDDEN, t);
-	}
-}
