--- src/main/java/com/workat/exception/dto/ErrorResponse.java
@@ -1,17 +1,0 @@
-package com.workat.exception.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@AllArgsConstructor
-@Data
-public class ErrorResponse {
-
-	private int code;
-
-	private String message;
-
-	public static ErrorResponse of(int code, String message) {
-		return new ErrorResponse(code, message);
-	}
-}
