+++ src/test/java/com/workat/domain/config/DataJpaTestConfig.java
@@ -1,0 +1,15 @@
+package com.workat.domain.config;
+
+import javax.persistence.EntityManager;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@TestConfiguration
+public class DataJpaTestConfig {
+
+	@Autowired
+	private EntityManager entityManager;
+}
